refactor(core): extract combo module, relocate inline tests, document combo/NC behavior
- move key-combo parsing + safety-block logic out of the press command into a
  neutral commands::combo module (registered as a non-command helper); key_down,
  key_up, press import from it
- relocate inline test blocks in action.rs, node.rs, interaction_policy.rs to
  sibling _tests.rs files, matching the project pattern
- document blocked-combo normalization + per-invocation limitation and the
  notification close-failure behavior in the skill command references
